Added Point Multiplier
Added the point multiplier upgrade! If you have enough points to afford it, clicking on it will give a 5x multiplier to all point gain.

+++ Assets/Scripts/Pachinko Scripts/SlotScript.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class SlotScript : MonoBehaviour
 {
     //Set in inspector. This is the base point value for each slot, before any multipliers are applied. 
-    public int basePoints;
+    public double basePoints;
 
     private void OnTriggerEnter(Collider other)
     {
